+++ scripts/test/Config/Defaults.t
@@ -32,15 +32,13 @@ my @ConfigFiles = $Kernel::OM->Get('Kernel::System::Main')->DirectoryRead(
     Filter    => "*.xml",
 );
 
+#rbo - T2016121190001552 - replaced list of XML files with KIX files
 my %AllowedConfigFiles = (
-    'CloudServices.xml'     => 1,
-    'Daemon.xml'            => 1,
+    'FAQ.xml'            	=> 1,
     'Framework.xml'         => 1,
-    'Fred.xml'              => 1,
     'GenericInterface.xml'  => 1,
-    'OTRSCodePolicy.xml'    => 1,
+    'ITSM.xml'    			=> 1,
     'ProcessManagement.xml' => 1,
-    'Support.xml'           => 1,
     'Ticket.xml'            => 1,
 );