+++ CHANGES.md
@@ -16,4 +16,4 @@
  * (2017/01/18) - Bugfix: T2017011390000391 (wrong column title in config item compare mask) (ddoerffel)
  * (2017/01/02) - CR: T2016122890000451 (added customer ticket template portal) (rbo)
  * (2016/12/23) - CR: T2016121990002948 (address address book functionality) (rbo)
- * (2016/12/06) - CR: T2016121190001552 (code merge of all packages) (ddoerffel)
+ * (2016/12/06) - CR: T2016121190001552 (code merge of all packages and changes for KIX 2017) (ddoerffel)